MB-36572: Set non-zero file descriptors in unit tests
When we start using caching file descriptors we need a non-zero limit
on the number that we can cache. Set the Environment to a non-zero
value in the unit tests and the mock server API so that we can set
cache limits correctly when that code is introduced.

We need to set this in both the mock server API and the unit tests. The
mock_server variant is used whenever we spin up an engine (i.e.
engine_testapp and SynchronousEPEngine) and the unit test variant is
required when we construct a file cache in the unit tests manually (i.e.
CouchKVStore tests).

Change-Id: I89a5756ec86eecd23cd9b40fcece73569ad3a51e
Reviewed-on: http://review.couchbase.org/c/kv_engine/+/134969
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>