MB-36564: Do not send first commit as commit when in disk snapshot
If we stream a disk snapshot to a replica we can send the first item
(if it is a commit and the replica already has the corresponding
prepare) as a Commit instead of as a Mutation. This should not cause
an issue on the replica as it already has the prepare and will not
attempt to move the HCS on disk, but it is added complexity that is
no longer needed (as we now send all commits as Mutations from disk
snapshots).

Change-Id: If0c70f87d81c7c0c4d520c0b7fcee3843da4e82d
Reviewed-on: http://review.couchbase.org/116736
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>