MB-36563: Move resolvedQ to trackedWrites before getting cookies
When we transition state away from active we need to notify any
clients of in-flight SyncWrites with ambiguous. This is necessary
to not keep around connections. We do this by calling
getCookiesForInFlightSyncWrites. This drains the resolvedQueue and
throws away the items though which is a bug as we need to put them
into trackedWrites so that we can transition to PDM if appropriate.
Without the items in trackedWrites we will throw exceptions as we
cannot find the correct SyncWrite to complete.

Change-Id: I9465db797ccf58b2693b60fc37cc54c8ba3417af
Reviewed-on: http://review.couchbase.org/116789
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>