Don't attempt to log if cb::logger not present
The cb::logger object should always be present in normal operation
(once logger initialiszed early in memcached startup, until just
before memcached terminates), however it _is_ possible in unit tests
or similar that we attempt to log a message when the logger doesn't
exist.

To ensure we don't crash in this situation, check for a non-null
logger object before attempting to dereference it.

Change-Id: I8d418e8b9903a4e2fcfc38fd713d17f522cd1cc6
Reviewed-on: http://review.couchbase.org/116855
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Trond Norbye <trond.norbye@couchbase.com>