MB-36631: Return ENGINE_TMPFAIL when failing to create sasl refresh thread
Previously we would disconnect ns_server if we failed to create
the thread to refresh the password database. ns_server don't like
that memcached is rude, and as a payback it'll kill memcached ;-)

This is a temporary failure, so we can might as well pass that
information on to ns_server.

Change-Id: I8dabbbc80c6013a798c6a64418e9396d0d6d3946
Reviewed-on: http://review.couchbase.org/116901
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>