MB-36554: Make sure the engine won't reallocate tracing vector
The memory accounting could get out of sync if the underlying
engine tried to add a trace element and the tracer needed to
reallocate memory in the vector. To work around this problem
make sure that there there is at least 10 free elements in
the trace array before calling into the engine.

Change-Id: I69144e14708bf4bcc5ee4d0c19e516cadecfebdf
Reviewed-on: http://review.couchbase.org/116928
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>