MB-36637: Avoid double-lock to vbstate at KVBucket:setVBucketState
With MB-36372 we always acquire an exclusive lock to vbstate at
vbstate-change. But, under certain conditions we may end up with
executing ActiveStream::setDead, which may lock to vbstate again.

To avoid the deadlock, with this patch KVBucket::setVBucketState
ends up calling an overloaded version of ActiveStream::setDead
that expects a lock to vbstate in input and avoid double-locking.

Change-Id: I52ec86f54ad36f7070865a45ff57d8d6cd10a2c1
Reviewed-on: http://review.couchbase.org/117006
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>