MB-36557: Avoid lock-inversion at set-vbstate and conn-disconnect
With MB-36372 we always acquire an exclusive lock to vbstate at
vbstate-change. Also, down in the call hierarchy we acquire the
connLock.
But, in the code executed by EPE::handleDisconnect we acquire the
two locks in opposite order.
TSAN spots the issue as soon as the two code paths are executed in
parallel, which is achieved by the test added with this patch.
To fix, with this patch we release the connLock at EPE::handleDisconnect
before we acquire the vbstateLock.

Change-Id: I8968acd5b1a285f93a4e37fa3c01bf6c776b3b66
Reviewed-on: http://review.couchbase.org/117017
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>