MB-36720: Don't advance HPS out of completed snapshot on replica
When I changed the HPS to move on non-durable writes I introduced a
bug where we can attempt to move the HPS out of a complete snapshot.
This will cause non-monotonic exceptions to fire and the replica to
expose an inconsistent HPS value. Correct this by not allowing it to
move the HPS value to one greater than the current snapshot end.

Change-Id: I0144b639a8990d75f1a79bdc36290f6ae06c3ada
Reviewed-on: http://review.couchbase.org/117206
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>