MB-36723: Optimize KVShard memory usage
Each KVShard owns a subset of the total vBuckets, based on the total
number of shards. For example, with 4 shards each KVShard owns up to
1024/4 = 256 vBuckets.

However, the current implementation of KVShard places its vBuckets in
a std::vector, with each element indexed by its VBid. As such, the
vector is sparse - for 4 shards only 1/4 of the vector will be
populated.

This is wasteful, however it becomes more significant when the number
of shards increase - for example with 16 shards there will be 16
KVShard objects, each with only 1/16th of their elements
populated.

Measuring the size of memory allocations (on Ubuntu 18.04) shows the
KVShard objects alone are 2MB, excluding any indirectly allocated
objects such as the actual VBuckets etc.

With this change that is reduced by a factor of num_shards.

Change-Id: I006e2e5378ea45b5ab47021553003d3c75b2da20
Reviewed-on: http://review.couchbase.org/117257
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>