MB-36744: Convert storage_age to microseconds
Currently we track storage_age, the time it takes from enqueueing
an Item til the flusher visits it, in seconds. This lacks the
granularity for any particularly meaningful comparisons. Convert
the queuedTime of an Item to track microseconds (this requires more
storage space so it has been moved to not increase the size of an
Item), and only set the time when doing normal mutations (queueDirty).

Change-Id: Ib0c72c9a9154a7efde3878ea46e50eb30427a2b5
Reviewed-on: http://review.couchbase.org/117317
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>