MB-36650: Ephemeral Replica resilient to unprepared Abort
In MB-35633 we started with processing (rather than ignoring) Abort
messages at Replica in the case where the related Prepare has never been
received. That may happen for both EP and Ephemeral bucket when Replica
receives a Backfill snapshot (as Active can deduplicate Prepare and
Abort).

With MB-35633 we cover two cases:
1) "key" does not exist in the HashTable when Replica receives the
  unprepared Abort
2) a nornal mutation for "key" is in the HashTable when the unprepared
  Abort is processed

This patch covers the third (and missing) case: a Completed (Committed /
Aborted) Prepare exists in the HashTable when the unprepared Abort is
processed, which can happen for Ephemeral.

Note that this patch adds full test-coverage for unprepared Abort for
both EP and Ephemeral. Some tests for EP spotted the issue in MB-36735
(EP bucket issue when the unprepared Abort is received and the Completed
Prepare is Level::PersistToMajority and is still tracked in the
PassiveDM). The tests covering that case are temporarily disabled.

Change-Id: I0287456c538ecbe7fe055b4ee91d2074eb815a1e
Reviewed-on: http://review.couchbase.org/117323
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>