MB-36735: EP Replica resilient to unprepared Abort
So far, at PDM we never expect to end-up with processing the completion
of a tracked Prepare twice. If that happens, PDM throws.

But, there is a case where that can happen:
1) PDM is tracking a Level::PersistToMajority completed (Committed /
  Aborted) Prepare (that may happen if the prepare is not
  locally-satisfied), and..
2) Replica starts receiving a disk-snapshot, and..
3) Replica receives an "unprepared abort" (possible only for
  disk-snapshot)

The above is perfectly legal, so Replica must be resilient to that
scenario.

Please refer to the MB description for full details.

Change-Id: I90c2238785296436837b37d19a53481e019372d6
Reviewed-on: http://review.couchbase.org/117380
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>