MB-36739: Rollback vbucket_state on commit failure
When commit fails, all items of the commit batch are transferred to the
reject queue and retried. However when the retry occurs, the
vbucket_state (in-memory) contains modifications which were made during
the failed flush, this leads to Expect failures, e.g. an pending Item's
seqno is now less or equal to the vbstate.highPreparedSeqno.

Primary change which fixes the crash is to roll back the vbucket_state
if the commit failed.

A second change is also added in which we remove the redundant
while(fail){sleep} logics from EPBucket::commit, there is no point in
spinning in a loop here because recalling KVStore::commit after failure
does nothing but return success, the failed KVStore::commit has moved
all items to the rejectQueue and cleared pendingReqs.

Finally a test is added which gives coverage for couchstore, a new
function is added to KVBucketTest (to allow other tests to pick up this
goodness) which allows the GMocked FileOpsInterface to be used to inject
failure into the flush.

Change-Id: Iee269664bd3682e4760674f4afe8ad7d4006bdb5
Reviewed-on: http://review.couchbase.org/117420
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>