MB-36784: Revert "MB-36227: Make EPStats cacheline padded"
This reverts commit dd7604da541a3003a62592944dec62be9a3cffab.

This commit appears to cause a regression as described in MB-36784. I
don't believe at this point in time that the regression is due to this
change, but it appears to exacerbate some other issue with these
performance tests that causes the results to vary by as much as 150%.

Initial perf runs of a revert of this against the build it was committed
in showed that the likelihood of test variance was diminished, albeit
still possible. Given that this patch didn't give us any notable
performance improvement, revert it for now to be revisited as part of
Cheshire-Cat.

Change-Id: I645e269ff07d633b9e9996395729fab7cfe8f4ff
Reviewed-on: http://review.couchbase.org/117493
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>