MB-56032 [2/4]: Set AuxIO threads to minimum priority
In patch 1 we increased the default number of AuxIO threads (for
machines with >8 CPU cores), to increase the potential IO queue depth
we can drive for Backfilling tasks and consequently increase the
throughput of Backfilling. However, we do _not_ want to adversly
impact latency-sensitve operations ruinning on other threads such as
front-end threads or Reader threads (which are part of the the
critical path of accessing non-resident docs).

To mitigate this, decrease the priority of AuxIO threads to lessen
their impact on other thread types.

See the original Writer thread MB-36723 and MB-36827 for more
background on the affect of (Writer) thread priority on front-end
threads.

Change-Id: I8a48dbcefb3307b56d1e9fed44ee2ea666ca267a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/192736
Well-Formed: Restriction Checker
Tested-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>