MB-33736: Create a larger value for mem stats test
Test can fail because when active compression is enabled the document
which is just 2,048 consecutive 'b' characters compresses very well. So
if the compressor manages to compress the value before eviction, the
observed mem_used change is tiny and doesn't take much jitter to make
the test fail.

Before this commit test failed often on CV and locally 30% failure rate
(for 23 runs0, after this commit 100% success (for 31 runs).

Change-Id: I72d1aef3f71ecacc446f49cd642028fa341e27bf
Reviewed-on: http://review.couchbase.org/117740
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>