MB-36794 Ensure we disconnect when we receive EINVAL
Ensure we disconnect as an producer when we receive EINVAL for DCP
{prepare, commit, abort} from the consumer. As currently we
just swallow these errors, which can cause the stream to deadlock.
As we may have thought we've sent all the data the consumer need but
in fact its view of the world is now logically incorrect. Thus,
disconnecting gives us an opportunity to re-connect and resolve this bad
state.

Also add a unit test to check that we return false in the correct
situations from DcpProducer::handleResponse(). Also fix a couple of
clang-tidy warnings.

Change-Id: Ia3709f6ec33b1f623a0ea77cf29b28e84a678c13
Reviewed-on: http://review.couchbase.org/117804
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>