MB-36765: Revert "MB-35308: Move HPS on non durable writes"
This reverts commit 1dcdf61e860ea8c18cb75144b049e0104ef31ef9.

Moving the HighPreparedSeqno (HPS) on non-durable writes introduced a
perf regression in non-durable workloads.

Reverting this patch restores the previous behaviour of moving the HPS
only when a Prepare becomes locally satisfied (e.g., a
PersistToMajority prepare is persisted).

ns_server will instead check both the HPS and HighSeqno to determine
which replica to promote as the HPS alone will no longer be sufficient.

This does reintroduce MB-36133, which was previously fixed as part of
the reverted patch. This will be resolved in a different way in a
following patch.

Change-Id: Iaabc514304997bbb39e406cfa159ed9e2ed5b859
Reviewed-on: http://review.couchbase.org/117869
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Dave Rigby <daver@couchbase.com>