MB-36860: Expand CheckpointTest to Ephemeral
While the initial purpose of this change was to expand the
CheckpointDurabilityTest fixture to Ephemeral, actually all the
CheckpointTest run only for EPBucket.

This patch contains the minimum change required for expanding the
fixture to Ephemeral. Some tests have been disabled for Ephemeral
as they require further work, which will be done in follow-up patches.

Change-Id: I286eaa52513fb8ee6b8e02f294e44c68dc991d29
Reviewed-on: http://review.couchbase.org/118119
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>