MB-36133: Persist highPreparedSeqno
If a replica crashes and warms up in the middle of a snapshot (i.e.,
the snapshot has only been partially persisted) it needs to warm up a
suitable HighPreparedSeqno. It cannot simply use the seqno of the most
recent prepare, as this prepare may be from a partial snapshot. The
in-memory snapshot is only allowed to move into complete snapshots to
avoid seqno acking prepares before the replica has reached a consistent
point (a snapshot end).

This patch adds a new field to vbstate `high_prepared_seqno`
which is the highest prepare seqno, from the most recent _fully
persisted_ snapshot.

Change-Id: I983866f0a11095a9e13ce1a808ae67e1ebb1a581
Reviewed-on: http://review.couchbase.org/117910
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>