MB-36904: Remove Flusher::State::Pausing
Remove the intermediate Flusher::State 'Pausing', advancing directly
from 'Running' -> 'Paused'.

Due to the changes in patch 1f64b6467 we no longer wake Flusher tasks
unless there's actual items to flush for that shard. As a consequence,
a Flusher now stays in State::Pausing until the next mutation arrives;
whereas prior to that patch it would typically get run "soon".

The flusher staying in state Pausing is benign from KV-Engine's pov,
however tools which change the flusher state (cbepctl) wait for the
state to change to 'Paused' to confirm their request was correct, and
as such now hang if Flusher task for shard 0 (primary shard) is idle
and has no items to flush.

Given the intermediate state 'Pausing' is unnecessary, remove it fix
this hang (and as a side-effect simplify the Flusher a little).

Change-Id: I5dad6c890e3e7924074fe8ad7fe77911e7502435
Reviewed-on: http://review.couchbase.org/117878
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>