MB-36794: Update highest deduped seqno on abort
We need to update the highest deduped seqno of the seqList when
we overwrite a prepare with an abort or we can end up streaming
an abort from memory that has de-duplicated the prepare before
it. This is an issue as we will fail sanity checks on the replica
and tear down the DCP connection.

Change-Id: I8487319652973da4a46bbd339b2f61e78e0ea1a4
Reviewed-on: http://review.couchbase.org/117890
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>