Cleanup: don't use extra buffering for stat response
It is already going into a buffer in bufferevents so we
don't need a continuous buffer to spool all of the stats
before sending it to bufferevents.

Previously we formatted one large buffer (and growing from
the add_stat method) and finally copied it over to
bufferevent / libevent once we had everything. Now that
we've moved over to bufferevent we can't add the stats
from _different_ threads (the executor pool) as we don't
hold the bufferevents lock while adding the buffer, causing
a thread sanitizer to warn when it release the buffer (now
holding the lock).

To work around this just keep the formatted stats in a
std::vector and pump them into the connections buffer
once we're done.

Change-Id: I5bc09d09ea4350178a19df0156b0f41f4492becb
Reviewed-on: http://review.couchbase.org/117904
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>