MB-36915: Avoid lock-inversion at set-vbstate and new-producer
With MB-36372 we always acquire an exclusive lock to vbstate at
vbstate-change. Also, down in the call hierarchy we acquire the
connLock.
But, in the code executed by EPE::dcpOpen we may acquire the
two locks in opposite order.
TSAN spots the issue as soon as the two code paths are executed in
parallel, which is achieved by the test added with this patch.
To fix, with this patch we release the connLock at EPE::dcpOpen
before we acquire the vbstateLock.

Change-Id: Idb5e0d9d837361a8f85beee2aa114fd9f1f85af0
Reviewed-on: http://review.couchbase.org/117947
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>