MB-23086: Disassociate ExecutorThread::run from any engine
The purpose of this commit is to ensure that the ExecutorThread doesn't
use a deleted engine. In this commit, the use of a deleted engine is
inert because it's just a pointer that is never dereferenced, however
that will change when ObjectRegistry needs the engine's arena.

To achieve this change the ExecutorThread::run so that is explicitly
allocates to no engine (using NonBucketAllocationGuard). Thus for the
majority of the run loop, ObjectRegistry stores null and only for a
small number of places should the thread place an engine in
ObjectRegistry.

1) When a task is invoked.

Switching to the engine is now done by using a new GlobalTask::execute
method which uses a BucketAllocationGuard to switch to the task's
engine. The execute method is now what calls the virtual run() method.

2) When a task is deleted.

In general tasks are deleted (actually destructed and deallocated) in
two places.

a) Within the ExecutorThread by calling ExecutorPool::cancel with
   eraseTask:true.
b) Within ~EventuallyPersistentEngine as the engine takes sole ownership
   of its tasks.

This commit changes case a) by using the BucketAllocationGuard around
the call to shared_ptr<GlobalTask>::reset(). Given the destruction of
an engine cannot occur until all tasks are destructed, it is safe to
switch to the engine before the task.reset(), after the reset the engine
can be deleted, but is no longer dereferenced.

Change-Id: Ied2ef3cb0adabe8634a3829260a4fc7562eedad5
Reviewed-on: http://review.couchbase.org/108349
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>