MB-36973: Don't use ThreadLocalPtr for CouchKVStore::pendingFileDeletions
CouchKVStore::pendingFileDeletions is a queue of pending files to be
deleted, which is currently implemented using AtomicQueue.

AtomicQueue internally uses a pthread_key to manage the
thread-specific data, however there is per-process limit on how many
pthread_key objects can be created - PTHREAD_KEYS_MAX. On current
Linux systems that is typically 1024.

The combination of a large bucket count and many shards can result in
us consuming all available pthread_key objects, resulting in bucket
creation failing with:

    WARNING 328 - Failed to create bucket [bucket-27]: Failed to create a thread-specific key: Resource temporarily unavailable: Resource temporarily unavailable

For example, with 32 buckets and 24 shards we will consumer 768
pthread_keys just for CouchKVStore objects, ignoring all other uses.

Fix this by removing the use of ThreadLocalPtr for
CouchKVStore::pendingFileDeletions - this is just a per-shard queue of
files ot delete, which is rarely going to be contended anyway so a
simple std::queue guarded via folly::Synchronized is sufficient.

Change-Id: Ie1d416630855aeed457d4db67b8b90f358b31357
Reviewed-on: http://review.couchbase.org/118197
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>