MB-36996: Replace remaining uses of AtomicQueue with folly Queue classes
ConnMap::pendingNotifications is implmented using our AtomicQueue
class. This internally uses ThreadLocalPtr, which in turn uses
pthread_key_create() to actually thread the thread-local.

As seen in MB-36948, there is a limit of how many pthread_key
thread-locals can be created - 1024 on Linux for example - and hence
using thread_key for things which are variable-sized at runtime is
potentially problematic.

As such, remove the remaining uses of AtomicQueue in favour of folly's
concurrent queue; which uses folly::ThreadLocal internally which
doesn't use a pthread_key per instance and hence avoids any problems
in hitting the limit.

Change-Id: I50dd6033fa161e0374d7163c3b60509a497a2e10
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/168449
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Richard de Mellow <richard.demellow@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>