MB-36971: Ensure that DCP Producer handles KeyEnoent correctly
Ensure that the DCP Producer disconnects when it receives a KeyEnoent
from a DCP consumer in response to a DcpPrepare, DcpCommit or DcpAbort.
As currently we swallow this error for these opcodes, which causes us to
hang in rebalance when the consumer has received a DcpCommit or
DcpAbort for a key it dose not have a prepare for. Disconnecting from
the consumer gives use the opportunity to re-connect and resolve the bad
state on the consumer.

We should also never see a KeyEnoent from a consumer in response to a
DcpPrepare. As when processing a prepare in KVBucket::prepare() we
always pass a 0 cas to VBucket::prepare() which will remap
MutationStatus::NotFound to ENGINE_KEY_ENOENT but the only cases where
processSet() and processSetInner() return NotFound are cases where a non
zero cas has been specified.

Also update a unit test to ensure we handle KeyEnoent and expand other
unit tests to make sure we handle Etmpfail and Enomem correctly.

Change-Id: Ie694f73c8e668298dd14d205ca753762cab7fab4
Reviewed-on: http://review.couchbase.org/118214
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>