MB-36948: Add vbucket_state::maxVisibleSeqno
Add a new seqno stored in the persisted vbucket state that tracks the
seqno of the most recent committed (via prepare or mutation) value.

This seqno will be used as the snapshot end when backfilling from disk
for a connection which is not sync-write aware.

Change-Id: Iee2f3d2688dfcaae530b675e14a3fc88d49da045
Reviewed-on: http://review.couchbase.org/118235
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>