MB-37505: Add Impersonate to mcbp protocol [1/2]
Impersonate allows a user to execute a command as another user without
authenticating as that user (similar to `sudo` in a Unix system). The
biggest difference between `impersonate` and the `sudo` command is
that the latter allows for privilege escalation. Impersonate will run
the command _as_ the requested user, but the effective privilege set
during the command execution is the intersection of the authenticated
users effective privilege set and the impersonated users available
privilege set.

Requirements:

* The connection must possess the `Impersonate` privilege
* The impersonated user must have access to Couchbase

Limitations:

* The impersonated user must be defined as a local user in Couchbase
  (fixing this require features in ns_server)
* The users with the `impersonate` privilege may execute commands as
  _any_ user defined in Couchbase (you can't specify that "Bob" can
  impersonate "Alice" but not "Joan")

Change-Id: I25ba3cd2b2e8cd5deb84b590f5a6c16d0a4bf4f7
Reviewed-on: http://review.couchbase.org/121070
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>