MB-36971: Send the HCS when streaming a Disk-Checkpoint
For Durability we need to keep track of the original source of a
Checkpoint, Disk/Memory. The reason is that, because of on-disk
deduplication and topology changes, a Producer we may end up with
streaming a disk-snapshot from memory. See MB-35245 for details.

We already send the HCS in the SnapshotMarker for disk-snapshot from
disk (done in MB-34017). We need to do the same for disk-snapshot from
memory.

By missing that we could end up with a newly promoted Active
re-committing already-committed Prepares and Replicas failing with
"Prepare not found in the HashTable" (legal at Replica, given that the
Prepare has been already committed and removed from the HT). See
MB-36971 for details.

Change-Id: I6e360cc81a6f05ea4165c82d754b3af1112bac5b
Reviewed-on: http://review.couchbase.org/118350
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>