MB-37013: Update DcpSnapShotMarker V2 to allow for an extra seqno
To address MB-36948 a new seqno is required in the snapshot marker.

This commit redefines the V2 format so that the V2 extras is a single
version field (1 byte). The 1 byte version field is used to describe how
the packet is to be decoded, in this MB one V2 encoding is created.

V2.0: The 'value' stores all of the snapshot marker data.
* start (u64)
* end (u64)
* flags (u32)
* maxVisibleSeqno (u64)
* highCompletedSeqno (u64)

With this commit only the V1 and V2.0 encodings are transmitted, but the
V2.0 encoding is only currently transmitted for snapshots with the disk
flag set (and when a sync write has occurred in the vbucket). Because
there is no path yet for the DcpProducer to push down the
maxVisibleSeqno, with this commit V2.0 always transmits 0 for the
maxVisibleSeqno.

On the receipt of a V2.0 message, this patch does tunnel the
maxVisibleSeqno up to KV-engine DCP, this is so the consumer can
calculate the correct number of bytes for buffer acknowledgement.

Note on removal of the following from dcp_snapshot_marker_executor.cc

  // HCS should never be sent as 0 or a pre-condition will throw in
  // the replicas flusher
  if (v2Payload->getHighCompletedSeqno() == 0) {
      // Not success so just disconnect
      connection.setState(StateMachine::State::closing);
      return;
  }

This disconnect clause is no longer required because analysis of the
ActiveStream::markDiskSnapshot shows that it would never push down a 0
HCS to the  Connection::marker send code.

Extended testing of the committed patch was performed using a multi
node cluster and various failover/rebalance steps to check the
disk-snapshot with the v2.0 encoding was transmitted and successfully
received.

Note: further testing was performed where the Connection::marker local
'maxVisibleSeqno' was initialised to a value to confirm that if all
snapshots used the V2.0 encoding, there is no issue.

Change-Id: I886503d6353d01b284b04af730d581f6be6784c9
Reviewed-on: http://review.couchbase.org/118387
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: James Harrison <james.harrison@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>