MB-36971: Never skip op::checkpoint_start in CM::getItemsForCursor
In MB-35245 we have introduced a bug where we may skip checkpoint_start
items in CheckpointManager::getItemsForCursor().
The side effect is that the MARKER_FLAG_CHK may not be set when sending
a SnapshotMarker to Replica. That may cause a Replica "merging"
snapshots of different types (Disk/Memory) into the same open
checkpoint.

Change-Id: Ie1ed7edef1914f365786ea4a70cccac1038f8350
Reviewed-on: http://review.couchbase.org/118409
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Ben Huddleston <ben.huddleston@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>