MB-36923: Fix test failures due to shard change
With the change to shard scaling by CPU we are now using more memory
out of the box on a system with a large number of CPUs. This is
causing issues with some memory sensitive tests. Drop the number of
shards back to 4 for these tests.

Some tests also specify a number of vBuckets. This must be greater
than the number of shards or tests will fail. Drop the number of
shards back to 4 for these tests.

The XDCR testsuite runs very close to the 60 second timeout for the
rockdb variant so bump the timeout to 120 seconds.

Change-Id: I718bfde05badfa7c15246b1f8d42a86288167fe9
Reviewed-on: http://review.couchbase.org/118781
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>