MB-37096: Debug CRT: Avoid spurious buffer overrun error in SerialisedDocKey
The SerialisedDocKey ctor calls std::copy() with the destination of
SerialisedDocKey::bytes, which is defined as `char[1]`. However, `bytes`
isn't actually a single element array - it is actually the first element
of a variable-length array allocated at the end of the fixed-size
SerialisedDocKey object.

However, the Windows Debug CRT doesn't know this and reports a buffer
overrun error.

Avoid the error by casting the array to a `char*`.

Change-Id: If1021e6cfc4e3ea181f1bd07ede61ae84e96d427
Reviewed-on: http://review.couchbase.org/119233
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>